requirements.txt
langchain==0.2.11
langchain_core==0.2.23
langchain_openai==0.1.17
langchain_pinecone==0.1.3
pandas==1.5.3
protobuf==5.27.2
pymongo==4.8.0
python-dotenv==1.0.1
google-cloud-storage==2.14.0